Add mpsc channels between nic and tcp
They are now running on their own threads, with communicating via
channels.

Controlling the TCP is done by acquiring a mutex on the control
structure.

At some point I'll probably look at crossbeam in order to use the
`select` macro. This will make it easier to use channels for the control
messages too.